Add xenstore_event and xsdev_event_one
Read device information out of xenstore.

Do some basic input validation for device_t's read from xenstore.  If a
field is not set, then we bail on adding the device.  We also ensure
that the number values are semi-valid.

It is okay for serial to be NULL, so that is non-fatal.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>