refactor udev_find_mode to return the type
udev_find_more and its helpers don't actually need the device_t
structure for any of their actions.  Rework the functions to return the
type which can then be assigned to device_t directly.

libusb_find_more_about_nic() needs to take the vid and pid and return an
integer type for integration with the other usbvm code.  Make the
change.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>