Call vusb_assign when a xenstore watch files
vusb_assign needs to run in usbvm, so call it when a xenstore watch
fires.  Check for an assign: prefix and handle those while logging
totally unexpected values.

watch_path is only set in dom0, so we can't access it in USB VM.  Move
its check after the watch_token check.  watch_token should only trigger
in dom0, and watch_path will be set in that case.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>