Run vusb_assign either locally or remotely
vusb-daemon dom0 must notify the usbvm to perform the vusb_assign.  We
use xenstore for the notification.

vusb_assign now needs a local and remote version to handle the different
cases.

xasprintf is no long static since it's called from vusb_assign_remote.

We need a 0 on success, so we have to convert the bool true of
xs_write to 0.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>