Add 'running' policy command
A 'running' rule will check for matching devices and VMs, but only apply
the rule to the first running VM found in policy search order.  It
will skip the assignment for shutdown VMs.

*This is a change from the previous first match only configuration.*

Prior to this new command, boot-time plugging devices required an
'always' or 'default' rule with a vm uuid.  However, that meant that
only the first entry in usb-rules would work for hotplug a device after
boot - a second VM could not be matched.

A generic 'allow' rule for a device would hotplug to any VM, but it
would not attach on boot.

The new rule will allow handling rules for the same vid:pid, but with
differing uuids.  The first match rule meant rule search would stop on
whichever VM came first.  Here we want to skip rules for non-running
VMs so the Nth VM can get the device assignment if it is running and the
others are not.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>