Populate devices from xenstore at startup
We could start up with devices already in xenstore.  While the xenstore
watch will fire, it'll be the top level directory which we don't
process.

Explicity iterate through the directory and grab any existing device
entries.  xs_directory gives us the node names and not the absolute
path, so we have to convert to absolute ourselves.

xsdev_fill should expanded out the node to the full path, so do that.
Also split xsdev_add out of xsdev_event_one and have xsdev_fill call
that directly.  There is no reason the fill function should possibly end
up in remove.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>