Prevent sticky & auto-assign of ambiguous devices
Add checks for device ambiguity to:
- Setting the sticky but
- Automatic assignment of new device
- Automatic assignment of created VM

This prevents policy rules from matching multiple devices, each of which
would be passed to a single VM instance. This is most common when
devices do not populate the serial number, thus there is not enough
information to uniquely identify based on policy rules. This does not
affect manual assignment, which uses USB bus and device IDs as opposed
to the vendor/product IDs of common rules such as the sticky/ALWAYS
mechanism.

OXT-1782

Signed-off-by: Kevin Pearson <kevin.pearson@ortmanconsulting.com>