Remove devices when usbvm shuts down
When the usbvm shuts down, thre is no clean up of the devices.  Current
code expects the /local/domain/Z/data/usb/devX-Y watch to be seen, but
for a shutdown, libxl (?) will just delete /local/domain/Z, which is
delivered as the watch_path /local/domain/Y/data/usb firing.  Therefore
we need to react to /local/domain/Y/data/usb firing by clearing out
devices.

This is simplistic for now - usbvm means it holds all devices and they
should all be removed.  Future enhancement with need to only remove
devices associated with a particular usbvm.

Note, there is something wrong with the UI updating on device removal.
dbus queries of vusb-daemon returns an empty list of devices, but the UI
is still showing devices from the shutdown usbvm.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>