xenstore: Lower "xenstore event" verbosity
This prints for the devX-Y node and all children, so 10 times per added
usb device.  That's too much, so make it debug so it won't be seen.

When a device is added, there is still an "xsdev_event_one adding device
X:Y" message.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
---
The multiple watch firings is unfortunate - we only need 1.  Maybe a
single xenstore node could contain all the data, but it is nicer to have
the key-value pairs in xenstore.

We could fill data one subtree, and have watches on a second.  When the
watch files, you have to go look in the other tree.  Splitting like that
isn't particularly appealing either.